livekit-examples/agent-demos/ivr-nagivator/requirements.txt
python-dotenv
flask
flask-cors
livekit
livekit-agents
livekit-plugins-silero
livekit-plugins-deepgram
livekit-plugins-openai
livekit-plugins-cartesia